requirements.txt
grpcio==1.30.0
grpcio-tools==1.29.0
numpy==1.19.0
protobuf==3.12.2
scipy==1.5.0